fixed loading bug
